Create Winmax4Service for API authentication
This update introduces a new Winmax4Service class for managing API authentication. The creation of this service allows us to adhere to the Single Responsibility Principle, as the Winmax4Controller now delegates API requests to this new service. This division of responsibilities results in a more manageable and clean codebase. Adjustments have also been made to the routing as part of this change.