Add functionality to regularly sync, manage and fetch Winmax4 families
A new model, migration and command for Winmax4Family has been created to manage 'Winmax4 families'. The syncFamilies command is added into the scheduling of commands in Winmax4ServiceProvider. The models now include the capability to sync and store families from Winmax4 API, and a new route has been set up in the controller to fetch these families.