Update debugging method in SyncFamiliesJob
The debugging method has been changed from 'dd' (die dump) to 'dump' in SyncFamiliesJob. This enables the continuation of the code during debugging and allows for a better understanding of the family structure update. The 'family_id' in the subfamily update/create method has been adjusted to use the ID of the new family for improved data consistency.