The provided commit message does not correspond to the provided diff. Based on the diff, a more accurate commit message could be:
Add code to examine server response in Winmax4Service

Added a line of code in Winmax4Service.php to simply dump the content of the response from the server. This allows developers to inspect the exact content received from the server before it is processed further by json_decode function.