Implement syncing entities with the Winmax4 API
A new job was added to synchronize entities with the Winmax4 API. This includes the creation of a new table 'winmax4_entities', a new model 'Winmax4Entity', and a new command 'syncEntities' with respective migration. A new method was also added to Winmax4Controller to facilitate access to these entities. Errors in the previous migrations 'create_winmax4_taxes_table' and 'create_winmax4_taxes_rates_table' have also been corrected.