Add license_id to fillable array in Winmax4Entity
The license_id field was added to the fillable array in the Winmax4Entity model. More importantly, the booted method was updated to add the LicenseScope only when the 'winmax4.use_license' config value is true and the app is not running in the console. This enhances flexibility by allowing the LicenseScope to be controlled via the application's configuration.