Comment out bindings in Winmax4ServiceProvider and constructor method in Winmax4CurrenciesController
This commit comments out a chunk of code that binds multiple controllers in Winmax4ServiceProvider.php and the constructor method of the Winmax4CurrenciesController class. Revisions like this are often done in the process of refactoring to improve the architecture of the system, typically to enable improvements to the testing strategy.