Extend classes in Controllers and uncomment method in Winmax4CurrenciesController
This commit extends the Controller class in multiple controllers and uncomment the constructor method in Winmax4CurrenciesController. It additionally comments out a block of code in Winmax4ServiceProvider that was previously binding different controllers. These changes are typically made during system refactoring to improve the code structure or the testing strategy.