Add delete functionality to Winmax4Entities
Implemented the functionality to delete entities from the Winmax4 API. This includes modifications to the Winmax4Service.php where the deletion logic is placed and, a new route in web.php file. Additionally, a new method has been introduced in the Winmax4EntitiesController.php for handling delete requests to the Winmax4 API service.