Refactor Winmax4Entity for conditional SoftDeletes
The commit changes the approach of applying the SoftDeletes trait in the Winmax4Entity class, according to the configuration variable. Instead of manually adding the trait, a global scope is added when the configuration is true. This gives better control over the deletion process. The ambiguous code for adding traits only if they don't exist is also removed.