Remove restore method and restoring event from ConditionalSoftDeletes
The restore method and the restoring event were removed from the ConditionalSoftDeletes trait. This results in more streamlined functionality of the trait as it uncouples the restoring events from the process of soft deleting. Refactoring was done to make sure the trait checks are in order and the 'deleting' event behavior is customized.